+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 -   Allow wildcards (`*`) in object titles (command `read`)
--   Print a list of available categories (command `listcategories`)
--   Print a list of available object types (command `listobjecttypes`)
+-   Print a list of available categories (command `categories`)
+-   Print a list of available object types (command `types`)
 -   Assign IPs to subnets (command `fixip`)
 -   Display error when user asks for specific category entries but category is not assigned to object type(s)